removed compare and diff tests from test_datachain
